🔨 [fix] #523 : 프로필 페이지 네트워크 병목 현상 완화
