+++ client/src/pages/ProfilePage/index.tsx
@@ -1,6 +1,5 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { useParams, useNavigate } from "react-router-dom";
-import { useQueryClient } from "react-query";
 import PageTemplate from "@pages/PageTemplate";
 import MyPageUserProfile from "@components/MyPageUserProfile";
 import CalendarHeatMap from "@components/CalendarHeatMap";
@@ -8,12 +7,11 @@ import RoutineScroller from "@components/RoutineScroller";
 import LogoutButton from "@components/LogoutButton";
 import NotificationButton from "@components/NotificationButton";
 import useUserInfo from "@hooks/query/user/useUserInfo";
-import { QueryKey, RoutePath } from "@constants/enums";
+import { RoutePath } from "@constants/enums";
 import { authStorage } from "src/services/ClientStorage";
 import * as s from "./style";
 
 const ProfilePage = () => {
-  const queryClient = useQueryClient();
   const navigate = useNavigate();
   const { userId } = useParams();
 
@@ -21,12 +19,6 @@ const ProfilePage = () => {
   const isOwner = profileUserId === authStorage.get();
 
   const { userInfo } = useUserInfo(profileUserId);
-
-  useEffect(() => {
-    if (isOwner) {
-      queryClient.invalidateQueries([QueryKey.USER_INFO, authStorage.get()]);
-    }
-  }, [isOwner, queryClient]);
 
   return (
     <PageTemplate isRoot={isOwner} title="프로필" topNavRightItem={<NotificationButton />}>
